example-image-processor/requirements.txt
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0